Use `i32` rather than `usize` as "default integer" in library template

`usize` was renamed from `uint` in order to convey that it's not the "default integer type". Guide new users to use `i32` instead of `usize`.

+++ src/cargo/ops/cargo_new.rs
@@ -876,7 +876,7 @@ fn main() {
 "
         } else {
             b"\
-pub fn add(left: usize, right: usize) -> usize {
+pub fn add(left: u64, right: u64) -> u64 {
     left + right
 }
